+++ views/adminlte/security/login_prestashop.php
@@ -128,7 +128,6 @@ $this->registerCss('
 
 		<?php $form = ActiveForm::begin([
 			'id' => 'login-form',
-			'enableCsrfValidation'   => true,
 			// AJAX validation burns single-use Turnstile tokens before the real submit
 			'enableAjaxValidation'   => !\cinghie\userextended\helpers\TurnstileVerifier::shouldProtectLogin(),
 			'enableClientValidation' => false,