+++ views/registration/connect.php
@@ -35,7 +35,6 @@ $this->params['breadcrumbs'][] = $this->title;
 
                 <?php $form = ActiveForm::begin([
                     'id' => 'connect-account-form',
-                    'enableCsrfValidation' => true,
                 ]); ?>
 
                 <?= $form->field($model, 'email') ?>